refactor: change app view structure
