+++ RSLang/src/index.html
@@ -8,7 +8,7 @@
     <title>RSLang</title>
 </head>
 
-<body class="body">
+<body class="body body_hidden">
     <div class="blackout blackout_hidden"></div>
     <div class="wrapper">
         <header id="header" class="header">
