+++ RSLang/src/index.ts
@@ -1,5 +1,5 @@
-import './style.scss';
 import App from './app';
+import './style.scss';
 
 const app = new App();
 app.start();
