+++ RSLang/src/sass/layout/_body.scss
@@ -1,0 +1,3 @@
+.body_hidden {
+    display: none;
+}
