+++ RSLang/src/view/handbook/levelCards/IlevelCards.ts
