+++ RSLang/src/view/handbook/templatesForElements/templateForCreatingNewElement.ts
