+++ RSLang/src/view/handbook/wordCardInfo/IwordCardInfo.ts
