+++ RSLang/src/view/handbook/wordCardInfo/wordCardInfo.ts
