+++ RSLang/src/view/handbook/wordCards/IwordCards.ts
