+++ RSLang/src/view/handbook/wordCards/wordCards.ts
