--- RSLang/src/view/views/handbook.ts
@@ -1,56 +1,0 @@
-export const sectionGames = `<section class="games content__item">
-                            <div class="game-panel">
-                                <a href="#" class="game-panel__link">Аудиовызов</a>
-                                <a href="#" class="game-panel__link">Спринт</a>
-                            </div>
-                        </section>`;
-
-export const sectionPaginationTop = `<section class="pagination-panel top content__item">
-                            <div class="pagination">
-                                <a id="prev" href="#" class="pagination__link">
-                                <img src="./public/assets/icons/arrow-left.svg" width="20" height="20" alt="prev">
-                                </a>
-                                <a id="first" href="#" class="pagination__link">1</a>
-                                <a id="page1" href="#" class="pagination__link">2</a>
-                                <a id="page2" href="#" class="pagination__link">3</a>
-                                <a id="page3" href="#" class="pagination__link">4</a>
-                                <a id="page4" href="#" class="pagination__link">5</a>
-                                <a id="page5" href="#" class="pagination__link">6</a>
-                                <a id="last" href="#" class="pagination__link">30</a>
-                                <a id="next" href="#" class="pagination__link">
-                                    <img src="./public/assets/icons/arrow-right.svg" width="20" height="20" alt="prev">
-                                </a>
-                            </div>
-                        </section>`;
-
-export const sectionDictionary = `<section class="dictionary-panel content__item">
-                            <div class="dictionary">
-                                <div class="dictionary__words"></div>
-                                <div class="dictionary__levels">
-                                    <a id="A1" href="#" class="dictionary__link">A1</a>
-                                    <a id="A2" href="#" class="dictionary__link">A2</a>
-                                    <a id="B1" href="#" class="dictionary__link">B1</a>
-                                    <a id="B2" href="#" class="dictionary__link">B2</a>
-                                    <a id="C1" href="#" class="dictionary__link">C1</a>
-                                    <a id="C2" href="#" class="dictionary__link">C2</a>
-                                </div>
-                            </div>
-                        </section>`;
-
-export const sectionPaginationBottom = `<section class="pagination-panel bottom content__item">
-                            <div class="pagination">
-                                <a id="prev-bottom" href="#" class="pagination__link">
-                                    <img src="./public/assets/icons/arrow-left.svg" width="20" height="20" alt="prev">
-                                </a>
-                                <a id="first-bottom" href="#" class="pagination__link">1</a>
-                                <a id="page1-bottom" href="#" class="pagination__link">2</a>
-                                <a id="page2-bottom" href="#" class="pagination__link">3</a>
-                                <a id="page3-bottom" href="#" class="pagination__link">4</a>
-                                <a id="page4-bottom" href="#" class="pagination__link">5</a>
-                                <a id="page5-bottom" href="#" class="pagination__link">6</a>
-                                <a id="last-bottom" href="#" class="pagination__link">30</a>
-                                <a id="next-bottom" href="#" class="pagination__link">
-                                    <img src="./public/assets/icons/arrow-right.svg" width="20" height="20" alt="prev">
-                                </a>
-                            </div>
-                        </section>`;
