--- RSLang/src/view/views/main.ts
@@ -1,98 +1,0 @@
-export const sectionAbout = `<section class="about content__item">
-                <h2 class="section-title about__title">RSlang</h2>
-                <p class="about__text">С нашим приложением учить английский язык легко и увлекательно! Учите слова
-                    эффективно с нашим словарём,
-                    разделённым на уровни сложности. Играйте в мини-игры учите, запоминайте и закрепляйте знания.
-                    Следите за вашей статистикой и вдохновляётесь результатами!</p>
-                <div class="about__image-wrap">
-                    <img src="./public/assets/images/about.png" width="50" height="50" alt="" class="about__image">
-                </div>
-                </section>`;
-
-export const sectionScope = `<section class="scope content__item">
-                <h2 class="section-title scope__title">Преимущества нашего приложения</h2>
-                <p class="scope__text">Мы разработали для вас удобные средства для скореёшего изучения языка и быстрого
-                    закрепления результата.</p>
-                <div class="scope-card scope__item">
-                    <img class="scope-card__img" src="./public/assets/icons/games.svg" alt="games">
-                    <h3 class="scope-card__title">Игры</h3>
-                    <p class="scope-card__text">Для изучения слов и усиления запоминания в приложении есть игры: Sprint
-                        и Audio Challenge,
-                        которые помогут вам "прокачать" свой словарный запас в игровой форме.</p>
-                </div>
-                <div class="scope-card scope__item">
-                    <img class="scope-card__img" src="./public/assets/icons/dictionary.svg" alt="handbook">
-                    <h3 class="scope-card__title">Продуманный учебник</h3>
-                    <p class="scope-card__text">Наш учебник состоит из шести разделов, разделённых по уровням сложности.
-                        Словарь представляет собой интерактивные карточки, содержащие перевод
-                        слова, тематическое изображение, произношение как слова отдельно, так и в составе фразы.</p>
-                </div>
-                <div class="scope-card scope__item">
-                    <img class="scope-card__img" src="./public/assets/icons/statistics.svg" alt="statistics">
-                    <h3 class="scope-card__title">Статистика</h3>
-                    <p class="scope-card__text">Весь ход обучения можно просмотреть в статистике, в которой представлены
-                        данные как за текущий день, так и за весь период
-                        обучения. Для вашего удобства информация представляется в виде таблицы или в виде графиков.</p>
-                </div>
-                <div class="scope-card scope__item">
-                    <img class="scope-card__img" src="./public/assets/icons/ui.svg" alt="interface">
-                    <h3 class="scope-card__title">Удобный интерфейс</h3>
-                    <p class="scope-card__text">Интерфейс приложения подстраивается под ваш уровень языка для облегчения
-                        процесса обучения.</p>
-                </div>
-            </section>`;
-
-export const sectionTeam = `<section class="team content__item">
-                <h2 class="section-title team__title">Наша команда</h2>
-                <div class="member-card team__item">
-                    <img src="./public/assets/icons/avatar-2.svg" alt="memeber photo" class="member-card__photo">
-                    <p class="member-card__role">Team leader</p>
-                    <p class="member-card__name">Татьяна</p>
-                    <p class="member-card__contribution">Разнообразный и богатый опыт постоянное
-                        информационно-пропогандистское обеспечение нашей деятельности играет важную роль в формировании
-                        системы массового участия. Для современного мира управление и развитие структуры способствует
-                        повышению качества модели развития. </p>
-                    <a href="#" class="member-card__github">
-                        <img class="member-card__github-img" src="./public/assets/icons/github.svg" alt="github">
-                        <span class="member-card__github-nickname">Github nick</span>
-                    </a>
-                </div>
-                <div class="member-card team__item">
-                    <img src="./public/assets/icons/avatar-1.svg" alt="memeber photo" class="member-card__photo">
-                    <p class="member-card__role">Developer</p>
-                    <p class="member-card__name">Артур</p>
-                    <p class="member-card__contribution">Задача организации, в особенности же постоянный количественный
-                        рост и сфера нашей активности играет важную роль в формировании дальнейших направлений развития.
-                        Таким образом социально-экономическое развитие требует анализа системы массового участия.</p>
-                    <a href="#" class="member-card__github">
-                        <img class="member-card__github-img" src="./public/assets/icons/github.svg" alt="github">
-                        <span class="member-card__github-nickname">Github nick</span>
-                    </a>
-                </div>
-                <div class="member-card team__item">
-                    <img src="./public/assets/icons/avatar-3.svg" alt="memeber photo" class="member-card__photo">
-                    <p class="member-card__role">Developer</p>
-                    <p class="member-card__name">Евгений</p>
-                    <p class="member-card__contribution">Таким образом дальнейшее развитие различных форм деятельности
-                        напрямую зависит от поэтапного и последовательного развития общества. Не вызывает сомнений, что
-                        рамки и место обучения кадров способствует повышению качества укрепления демократической
-                        системы.</p>
-                    <a href="#" class="member-card__github">
-                        <img class="member-card__github-img" src="./public/assets/icons/github.svg" alt="github">
-                        <span class="member-card__github-nickname">Github nick</span>
-                    </a>
-                </div>
-                <div class="member-card team__item">
-                    <img src="./public/assets/icons/avatar-4.svg" alt="memeber photo" class="member-card__photo">
-                    <p class="member-card__role">Mentor</p>
-                    <p class="member-card__name">Дмитрий</p>
-                    <p class="member-card__contribution">Задача организации, в особенности же дальнейшее развитие
-                        различных форм деятельности требует от нас анализа укрепления демократической системы. Не
-                        вызывает сомнений, что новая модель организационной деятельности позволяет выполнять важные
-                        задания по разработке поэтапного и последовательного развития общества. </p>
-                    <a href="#" class="member-card__github">
-                        <img class="member-card__github-img" src="./public/assets/icons/github.svg" alt="github">
-                        <span class="member-card__github-nickname">Github nick</span>
-                    </a>
-                </div>
-                </section>`;
