--- RSLang/src/view/views/popup.ts
@@ -1,14 +1,0 @@
-export const popupMessage = `
-                <a href="#header" class="popup__area"></a>
-                <div class="popup__body">
-                    <div class="popup__content">
-                        <a href="#header" class="popup__close">
-                            <img src="./public/assets/icons/cross.svg" alt="close">
-                        </a>
-                        <div class="popup__title">
-                            <h2 id="popup-title" class="popup__title-text">Заголовок попап</h2>
-                        </div>
-                        <p class="popup__message">Не вызывает сомнений, что курс на социально-ориентированный национальный проект напрямую зависит от поэтапного и последовательного развития общества. Следует отметить, что социально-экономическое развитие напрямую зависит от дальнейших направлений развития. Значимость этих проблем настолько очевидна, что реализация намеченных плановых заданий способствует повышению качества экономической целесообразности принимаемых изменений. Для современного мира социально-экономическое развитие играет важную роль в формировании прогресса профессионального общества. Идейные соображения высшего порядка, а также курс на социально-ориентированный национальный проект позволяет выполнять важные задания по разработке системы массового участия. Прежде всего начало повседневной работы по формированию позиции обеспечивает широкому кругу специалистов направлений прогрессивного развития. Прежде всего консультация с широким активом напрямую зависит от системы массового участия. Не вызывает сомнений, что новая модель организационной деятельности напрямую зависит от поставленных обществом и правительством задач. Прежде всего новая модель организационной деятельности напрямую зависит от системы массового участия. Для современного мира рамки и место обучения кадров позволяет выполнять важные задания по разработке направлений прогрессивного развития.</p>
-                        <button class="popup__close-btn">Закрыть</button>
-                    </div>
-                </div>`;
