--- RSLang/src/view/views/viewType.ts
@@ -1,4 +1,0 @@
-export interface ViewType {
-  sections: string[]
-  name: string;
-}
