+++ RSLang/src/view/viewsContent/gameschoice.ts
@@ -1,0 +1,32 @@
+const sectionGamesChoice = `<section class="choice-panel content__item">
+                <div class="choice">
+                    <h2 class="choice__title">Мини игры</h2>
+                    <div class="choice__wrap sprint">
+                        <h3 class="choice__name">Спринт</h3>
+                        <img src="./public/assets/images/sprint.png" width="40" height="40" alt="" class="choice__img">
+                        <button class="choice__btn" id="sprint-game">
+                            <img src="./public/assets/icons/timer.svg" width="50" height="50" alt="">
+                        </button>
+                    </div>
+                    <div class="choice__wrap audio">
+                        <h3 class="choice__name">Аудиовызов</h3>
+                        <img src="./public/assets/images/audio.png" width="40" height="40" alt="" class="choice__img">
+                        <button class="choice__btn" id="audio-game">
+                            <img src="./public/assets/icons/sound.svg" width="50" height="50" alt="">
+                        </button>
+                    </div>
+                    <div class="choice__levels">
+                        <a id="A1" href="#" class="choice__level">A1</a>
+                        <a id="A2" href="#" class="choice__level">A2</a>
+                        <a id="B1" href="#" class="choice__level">B1</a>
+                        <a id="B2" href="#" class="choice__level">B2</a>
+                        <a id="C1" href="#" class="choice__level">C1</a>
+                        <a id="C2" href="#" class="choice__level">C2</a>
+                    </div>
+                    <div class="choice__hint">
+                        <p>Выберите уровень сложности</p>
+                    </div>
+                </div>
+            </section>`;
+
+export default sectionGamesChoice;
