+++ RSLang/src/view/viewsContent/sprintgame.ts
@@ -1,0 +1,20 @@
+export const sectionSprintGameStart = `
+            <section class="sprint-game content__item">
+                <div class="sprint-game__about">
+                    <h2 class="sprint-game__title">Спринт</h2>
+                    <p class="sprint-game__desc">Используйте эту тренировку для ...</p>
+                    <img class="sprint-game__icon" src="./public/assets/images/sprint.png" width="300" height="150" alt="">
+                </div>
+                <div class="sprint-game__help">
+                    <p class="sprint-game__help-item">Текст помощи</p>
+                    <p class="sprint-game__help-item">Текст помощи</p>
+                    <p class="sprint-game__help-item">Текст помощи</p>
+                    <p class="sprint-game__help-item">Текст помощи</p>
+                    <p class="sprint-game__help-item">Текст помощи</p>
+                </div>
+                <div class="sprint-game__start">
+                    <button class="sprint-game__start-btn">Начнём!</button>
+                </div>
+            </section>`;
+
+export const sectionSprintGameBody = '';
