+++ RSLang/src/view/viewsContent/viewType.ts
@@ -1,0 +1,4 @@
+export interface ViewType {
+  sections: string[]
+  name: string;
+}
